+++ jetty-bom/pom.xml
@@ -345,11 +345,6 @@
         <version>9.4.31-SNAPSHOT</version>
       </dependency>
       <dependency>
-        <groupId>org.eclipse.jetty</groupId>
-        <artifactId>jetty-ssl-reload</artifactId>
-        <version>9.4.31-SNAPSHOT</version>
-      </dependency>
-      <dependency>
         <groupId>org.eclipse.jetty.websocket</groupId>
         <artifactId>javax-websocket-client-impl</artifactId>
         <version>9.4.31-SNAPSHOT</version>
