+++ jetty-server/src/main/config/etc/jetty-ssl-reload.xml
@@ -4,7 +4,7 @@
 
   <Call name="addBean">
     <Arg>
-      <New id="sslKeyStoreScanner" class="org.eclipse.jetty.ssl.reload.SslKeyStoreScanner">
+      <New id="sslKeyStoreScanner" class="org.eclipse.jetty.util.ssl.SslKeyStoreScanner">
         <Arg><Ref refid="sslContextFactory"/></Arg>
         <Set name="scanInterval"><Property name="jetty.ssl.reload.scanInterval" default="1"/></Set>
       </New>
