--- jetty-ssl-reload/src/test/resources/jetty-logging.properties
@@ -1,4 +1,0 @@
-org.eclipse.jetty.util.log.class=org.eclipse.jetty.util.log.StdErrLog
-org.eclipse.jetty.deploy.LEVEL=WARN
-org.eclipse.jetty.util.Scanner=WARN
-org.eclipse.jetty.ssl.reload.LEVEL=DEBUG
