+++ jetty-util/src/main/java/org/eclipse/jetty/util/ssl/SslKeyStoreScanner.java
@@ -16,7 +16,7 @@
 //  ========================================================================
 //
 
-package org.eclipse.jetty.ssl.reload;
+package org.eclipse.jetty.util.ssl;
 
 import java.io.File;
 import java.net.URI;
@@ -29,7 +29,6 @@ import org.eclipse.jetty.util.annotation.ManagedOperation;
 import org.eclipse.jetty.util.component.AbstractLifeCycle;
 import org.eclipse.jetty.util.log.Log;
 import org.eclipse.jetty.util.log.Logger;
-import org.eclipse.jetty.util.ssl.SslContextFactory;
 
 public class SslKeyStoreScanner extends AbstractLifeCycle implements Scanner.DiscreteListener
 {
