+++ pom.xml
@@ -95,7 +95,6 @@
     <module>jetty-server</module>
     <module>jetty-xml</module>
     <module>jetty-security</module>
-    <module>jetty-ssl-reload</module>
     <module>jetty-openid</module>
     <module>jetty-servlet</module>
     <module>jetty-webapp</module>
