+++ tests/test-integration/src/test/resources/badKeystore
