+++ tests/test-integration/src/test/resources/newKeystore
