+++ tests/test-integration/src/test/resources/oldKeystore
